// src/app/services/api/model/inscripcion.ts
/**
 * ApiEventos
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Inscripcion { 
    idInscripcion?: number;
    idEvento?: number;
    idUsuario?: number;
    fechaInscripcion?: string;
}